scripts/requirements.txt
# ─────────────────────────────────────────────
# ML Pipeline: STT + LLM Fine-tuning
# Python 3.10 tavsiya etiladi
# ─────────────────────────────────────────────

# ── PyTorch (CUDA 11.8 bilan o'rnatish) ──
# pip install torch torchaudio --index-url https://download.pytorch.org/whl/cu118
torch>=2.1.0
torchaudio>=2.1.0

# ── Transformers & Training ──
transformers>=4.40.0
datasets>=2.18.0
accelerate>=0.27.0
evaluate>=0.4.1

# ── STT - Phase 1 ──
librosa>=0.10.0
soundfile>=0.12.1
jiwer>=3.0.3

# ── LLM Summarization - Phase 2 ──
sentencepiece>=0.1.99
rouge_score>=0.1.2

# ── Uzbek Dataset yaratish (Gemini API) ──
google-genai>=1.0.0

# ── Umumiy ──
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
